Replace use of MathJax with MathML Core
MathML Core is supported by Chromium since the start of 2023, and Firefox and Safari have supported MathML for many years.

While the markup is a lot more difficult to read than LaTeX markup, the output is very similar, I don't use it that much, and it saves having to download ~ 1.2 MB of MathJax data to view a page that uses it, so the pages are about 4x lighter.