Vendor Mermaid
Also update it to the latest release.

Vendoring Mermaid is a bit of a pain, because while there is a single minified JS file available, it's 3.2 MB and includes a lot that this site doesn't need or download from the CDN when using ES modules.

As such, this vendors the ES modules that this site loads, but only those files, which means that future use of Mermaid that uses other functionality may need more modules vendored.